Implement news article & folders refactor
