+++ .gitignore
@@ -228,6 +228,5 @@ _data.json
 session_data
 assets
 node_modules
-src/infrastructure/static/img/news
-src/infrastructure/db/storage.json
-storage.json
+src/core/infrastructure/db/storage.json
+src/core/infrastructure/db/news_article_images/*
