+++ pyproject.toml
@@ -39,7 +39,7 @@ lint.ignore = [
   "DJ008",
   "PTH123",
   "E731",
-  "C408",   
+  "C408",
   "TD002",
   "TD003",
   "FIX002",
