--- src/application/api/index.py
@@ -1,29 +1,0 @@
-"""Index router."""
-
-from litestar import Router, get
-from litestar.contrib.htmx.response import HTMXTemplate
-from tinydb import TinyDB
-
-
-from src.application.repository import NewsRepository
-from src.infrastructure.db import get_db_session
-
-
-@get("/", name="index", exclude_from_auth=True)
-async def index(db_session: TinyDB) -> HTMXTemplate:
-    """Index page."""
-    news_repository = NewsRepository(db_session)
-
-    return HTMXTemplate(
-        template_name="index.html",
-        context={"articles": news_repository.get_all()[:2]},
-    )
-
-
-router = Router(
-    route_handlers=[
-        index,
-    ],
-    path="/",
-    dependencies={"db_session": get_db_session},
-)
