+++ src/core/domain/news.py
@@ -1,0 +1,9 @@
+from src.core.infrastructure.types import Entity
+
+
+class NewsArticle(Entity):
+    """Domain model for news articles."""
+
+    title: str
+    description: str
+    images: list[str]
