+++ src/core/infrastructure/db/__init__.py
@@ -1,0 +1,5 @@
+from tinydb import TinyDB
+
+
+class DBSession(TinyDB):
+    pass
