--- src/infrastructure/templates/auth/login.html
@@ -1,17 +1,0 @@
-{% extends "shared/base.html" %}
-{% block content %}
-
-
-<div class="uk-section uk-flex uk-flex-middle uk-flex-center uk-height-viewport">
-  <div class="uk-card uk-card-default uk-card-body uk-width-1-3@m">
-    <h3>Log in</h3>
-    <form hx-post="login_form" hx-target="#message">
-      <div class="uk-margin"><input type="text" name="username" placeholder="Username" required></div>
-      <div class="uk-margin"><input type="password" name="password" placeholder="Password" required></div>
-      <div class="uk-margin"><button type="submit">Login</button></div>
-    </form>
-    <div id="message"></div>
-  </div>
-</div>
-
-{% endblock %}
