--- src/infrastructure/templates/index.html
@@ -1,80 +1,0 @@
-{% extends "shared/base.html" %}
-{% block content %}
-
-<!-- HERO -->
-<section class="uk-section uk-section-small">
-  <div class="uk-container">
-    <div class="uk-height-large uk-cover-container uk-border-rounded">
-      <img src="{{ url_for("static", file_path="img/home.jpg" ) }}" alt="Alt img" data-uk-cover>
-      
-    </div>
-  </div>
-</section>
-<!-- HERO -->
-
-<!--CONTENT-->
-<div class="uk-section uk-section-default">
-  <div class="uk-container">
-    <div class="uk-grid" data-ukgrid>
-      <div class="uk-width-2-3@m">
-        <h4 class="uk-heading-line uk-text-bold"><span>Новости</span></h4>
-        {% for article in articles %}
-
-          <article class="uk-section uk-section-small uk-padding-remove-top">
-
-            <header>
-              <h2 class="uk-margin-remove-adjacent uk-text-bold uk-margin-small-bottom">
-                <a title="Fusce facilisis tempus magna ac dignissim." class="uk-link-reset"
-                  href="#">
-                  {{ article.title }}
-                </a>
-              </h2>
-            </header>
-            {% for path in article.images_path %}
-              <figure>
-                <img 
-                    src="{{ url_for("static", file_path=path|safe ) }}"
-                  width="800" height="300"
-                  alt="Alt text" class="lazy" data-uk-img>
-                <figcaption class="uk-padding-small uk-text-center uk-text-muted">Caption of the image</figcaption>
-              </figure>
-
-            {% endfor %}
-
-            {{ article.description|safe }} 
-
-            <h4 class="uk-heading-line uk-text-bold"><span></span></h4>
-          </article>
-        
-        {% endfor %}
-      </div>
-
-      <div class="uk-width-1-3@m">
-        <div data-uk-sticky="offset: 90;media:@m">
-
-          <h4 class="uk-heading-line uk-text-bold"><span>Archive</span></h4>
-          <ul class="uk-list">
-            <li><a href="http://gosuslugi.ru/">Госуслуги</a></li>
-            <li><a href="http://gosuslugi.ru/">Госуслуги</a></li>
-            <li><a href="http://gosuslugi.ru/">Госуслуги</a></li>
-            <li><a href="http://gosuslugi.ru/">Госуслуги</a></li>
-          </ul>
-
-          <h4 class="uk-heading-line uk-text-bold"><span>Контакты</span></h4>
-          <div class="uk-tile uk-tile-small uk-tile-muted uk-border-rounded">
-                <p>ул. Сибиряков-Гвардейцев, 56/1</p>
-                <p>Торговый город "Левобережный"</p>
-                <p>Пн. - Сб. с 8.00 по 18.00</p>
-                <p>(383)342-82-96</p>
-                <p><a href="mailto:ncpo_tr@edu54.ru">ncpo_tr@edu54.ru</a></p>
-                <p><a href="http://центр54.рф" target="_blank">центр54.рф</a></p>
-          </div>
-        </div>
-      </div>
-
-    </div>
-  </div>
-</div>
-<!--/CONTENT-->
-
-{% endblock %}
