+++ src/main.py
@@ -1,6 +1,6 @@
 """Main entry point for the application."""
 
-from src.application.app import create_app
-from src.infrastructure.settings import app_settings
+from src.core.infrastructure.app import create_app
+from src.core.infrastructure.settings import app_settings
 
 app = create_app(app_settings=app_settings)
