+++ src/templates/shared/base.html
@@ -1,0 +1,44 @@
+<!doctype html>
+<html class="h-full bg-white">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  
+
+  <link href="{{ url_for("static", file_path="css/header.css" ) }}" rel="stylesheet" />
+  <link href="{{ url_for("static", file_path="assets/css/output.css" ) }}" rel="stylesheet" />
+
+  <script src="{{ url_for("static", file_path="htmx.org/dist/htmx.min.js" ) }}"></script>
+  <script src="{{ url_for("static", file_path="alpinejs/dist/cdn.min.js" ) }}"></script>
+
+  <script src="https://unpkg.com/hyperscript.org@0.9.12"></script>
+  <script src="https://unpkg.com/htmx-ext-class-tools@2.0.0/class-tools.js"></script>
+
+  <!-- UIkit CSS -->
+  <link rel="stylesheet" href="{{ url_for("static", file_path="uikit/dist/css/uikit.min.css" ) }}" />
+
+  <!-- UIkit JS -->
+  <script src="{{ url_for("static", file_path="uikit/dist/js/uikit.min.js" ) }}"></script>
+  <script src="{{ url_for("static", file_path="uikit/dist/js/uikit-icons.min.js" ) }}"></script>
+  
+    
+  <meta name="htmx-config" content='{"responseHandling": [{"code":".*", "swap": true}]}' /> <!--all responses are swapped-->
+
+  <title>{% block title %} Education Organization {% endblock %}</title>
+</head>
+
+<body class="h-full">
+  {% block header %}
+  {% include "/shared/header.html" %}
+  {% endblock %}
+
+  {% block content %}
+  {% endblock %}
+
+  {% block footer %}
+  {% include "/shared/footer.html" %}
+  {% endblock %}
+</body>
+
+</html>
