+++ src/templates/shared/header.html
@@ -1,0 +1,38 @@
+<header id="header" style="background-color: #fff;"
+  data-uk-sticky="show-on-up: true; animation: uk-animation-fade; media: @l">
+  <div class="uk-container">
+    <nav id="navbar" data-uk-navbar="mode: click;">
+      <div class="uk-navbar-center nav-overlay">
+        <a class="uk-navbar-item uk-logo" href="/" title="Logo">
+          <img src="{{ url_for("static", file_path="img/logo.png" ) }}" alt="Logo">
+        </a>
+      </div>
+      <div class="uk-navbar-right nav-overlay">
+        <a class="uk-navbar-item" href=/add_news_article>
+          {% if 'user_id' in request.session %}
+              Создать новость
+          {% endif %}
+        </a>
+      </div>
+    </nav>
+    <nav id="navbar" data-uk-navbar="mode: click;">
+      <div class="uk-navbar-center nav-overlay">
+        ГБПОУ НСО «Новосибирский центр профессионального обучения в сфере транспорта»
+      </div>
+    </nav>
+  </div>
+</header>
+
+<!-- NAVIGATION -->
+<div class="uk-container">
+  <nav class="nav-scroll">
+    <ul class="uk-subnav uk-flex uk-flex-between uk-flex-nowrap">
+      <li><a href="#">Основные сведения об образовательной организации</a></li>
+      <li><a href="#">Студентам и выпускникам</a></li>
+      <li><a href="#">Абитуриентам</a></li>
+      <li><a href="#">Работодателям</a></li>
+      <li><a href="#">Сотрудникам</a></li>
+    </ul>
+  </nav>
+</div>
+<!-- /NAVIGATION -->
